backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
openai
python-dotenv
requests
mcp